Add separate mod file (and workspace file)
