+++ exp/go.mod
@@ -1,0 +1,12 @@
+module github.com/prometheus/client_golang/exp
+
+go 1.21
+
+require (
+	github.com/google/go-cmp v0.6.0
+	github.com/klauspost/compress v1.17.7
+	github.com/prometheus/common v0.48.0
+	google.golang.org/protobuf v1.33.0
+)
+
+require github.com/prometheus/client_model v0.5.0 // indirect
